+++ apsw/tests/ftstests.py
@@ -367,6 +367,10 @@ class FTS(unittest.TestCase):
                     ):
                         self.db.fts5_tokenizer(tokenizer, args)(text.encode(), reason_code, None)
 
+    def testStr(self):
+        foo = self.db.fts5_tokenizer("ascii")
+        self.assertIsNotNone(re.match(r'<apsw.FTS5Tokenizer \"ascii\" args \(\) at 0x[a-f0-9]+>', str(foo)))
+
     def testAPSWFTSTokenizers(self):
         "Test apsw.fts5 tokenizers"